% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_clean.R
\name{addLabelDf}
\alias{addLabelDf}
\title{Add value labels to the data.frame}
\usage{
addLabelDf(data, mapping = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{mapping}{Set of aesthetic mappings created by aes or aes_.}
}
\description{
Add value labels to the data.frame
}
