% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{browsers}
\alias{browsers}
\title{Browser market share 2011}
\format{A data.frame with 12 rows and 3 columns
\describe{
    \item{browser}{browser}
    \item{version}{browser version}
    \item{share}{market share, in percentage}
}}
\usage{
browsers
}
\description{
A phony dataset measuring browser market share
}
\keyword{datasets}

