% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRadar.R
\name{coord_radar}
\alias{coord_radar}
\title{The radar coordinate system is a modification of polar coordinate system, commonly used for radar chart}
\usage{
coord_radar(theta = "x", start = 0, direction = 1)
}
\arguments{
\item{theta}{variable to map angle to (x or y)}

\item{start}{offset of starting point from 12 o'clock in radians}

\item{direction}{1, clockwise; -1, counterclockwise}
}
\description{
The radar coordinate system is a modification of polar coordinate system, commonly used for radar chart
}
