% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggArea.R
\name{ggArea}
\alias{ggArea}
\title{Draw an interactive area plot}
\usage{
ggArea(
  data,
  mapping,
  position = "stack",
  palette = "Blues",
  reverse = TRUE,
  alpha = 0.4,
  size = 0.3,
  use.label = TRUE,
  use.labels = TRUE
)
}
\arguments{
\item{data}{A data.frame}

\item{mapping}{Set of aesthetic mappings created by aes or aes_.}

\item{position}{Either "stack" or "fill"}

\item{palette}{A character string indicating the color palette}

\item{reverse}{If true, reverse palette colors}

\item{alpha}{Transparency}

\item{size}{Line size}

\item{use.label}{Logical. Whether or not use column label in case of labelled data}

\item{use.labels}{Logical. Whether or not use value labels in case of labelled data}
}
\value{
An area plot
}
\description{
Draw an interactive area plot
}
\examples{
require(gcookbook)
require(ggplot2)
ggArea(uspopage,aes(x=Year,y=Thousands,fill=AgeGroup))
ggArea(uspopage,aes(x=Year,y=Thousands,fill=AgeGroup),position="fill")
}
