% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCatepillar.R
\name{ggCatepillar}
\alias{ggCatepillar}
\title{Make an interactive catepillar plot}
\usage{
ggCatepillar(
  data,
  mapping,
  errorbar = "se",
  interactive = FALSE,
  digits = 1,
  flip = FALSE,
  use.label = TRUE,
  use.labels = TRUE
)
}
\arguments{
\item{data}{a data.frame}

\item{mapping}{Set of aesthetic mappings created by aes or aes_.}

\item{errorbar}{which value is displayed with errorbar :"se" or "sd"}

\item{interactive}{A logical value. If TRUE, an interactive plot will be returned}

\item{digits}{An integer indicating the number of decimal places}

\item{flip}{Logical. If TRUE, coord_flip() function is used to make a horizontal plot}

\item{use.label}{Logical. Whether or not use column label in case of labelled data}

\item{use.labels}{Logical. Whether or not use value labels in case of labelled data}
}
\value{
An interactive catepillar plot
}
\description{
Make an interactive catepillar plot
}
\examples{
require(moonBook)
require(ggiraph)
require(ggplot2)
ggCatepillar(acs,aes(Dx,age,color=HBP))
ggCatepillar(acs,aes(c(Dx,sex),age,color=HBP),interactive=TRUE,flip=TRUE,use.labels=FALSE)
ggCatepillar(acs,aes(age,height,color=sex),errorbar=FALSE,interactive=TRUE)
}
