% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRose.R
\name{ggRose}
\alias{ggRose}
\title{Draw an interactive Rose plot}
\usage{
ggRose(data, mapping, palette = "Reds", color = "black", size = 0.1, ...)
}
\arguments{
\item{data}{A data.frame}

\item{mapping}{Set of aesthetic mappings created by aes or aes_.}

\item{palette}{A character string indicating the color palette}

\item{color}{Bar colour}

\item{size}{Bar size}

\item{...}{other arguments passed on to geom_bar_interactive.}
}
\value{
An interactive Rose plot
}
\description{
Draw an interactive Rose plot
}
\examples{
require(moonBook)
require(ggplot2)
require(ggiraph)
require(plyr)
ggRose(rose,aes(x=Month,fill=group,y=value),interactive=TRUE)
ggRose(acs,aes(x=Dx,fill=smoking),interactive=TRUE)
}

