% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPoints.R
\name{makeEq}
\alias{makeEq}
\title{Make a regression equation of a model}
\usage{
makeEq(model, digits = 2)
}
\arguments{
\item{model}{A model of class "lm" or"glm" or"loess"}

\item{digits}{integer indicating the number of decimal places}
}
\description{
Make a regression equation of a model
}
