% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPoints.R
\name{model2df}
\alias{model2df}
\title{Make a data.fram of yhat with a model}
\usage{
model2df(model, x = NULL, n = 100)
}
\arguments{
\item{model}{A model of class "lm" or"glm" or"loess"}

\item{x}{A optional vector of explanatory variable}

\item{n}{number of observations.}
}
\description{
Make a data.fram of yhat with a model
}

