% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPair.R
\name{myscale}
\alias{myscale}
\title{Rescale a vector with which minimum value 0 and maximum value 1}
\usage{
myscale(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
Rescale a vector with which minimum value 0 and maximum value 1
}
