% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPair.R
\name{myscale2}
\alias{myscale2}
\title{Rescale a vector with which minimum value 0 and maximum value 1}
\usage{
myscale2(x, minx = 0, maxx = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{minx}{The minimum value}

\item{maxx}{The maximum value}
}
\description{
Rescale a vector with which minimum value 0 and maximum value 1
}
