% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict3d.R
\name{myseq}
\alias{myseq}
\title{Generate regular sequences of desired length between minimum and maximal values}
\usage{
myseq(x, length = 20)
}
\arguments{
\item{x}{a numeric vector}

\item{length}{desired length of the sequence}
}
\description{
Generate regular sequences of desired length between minimum and maximal values
}

