% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggSpine.R
\name{num2cut}
\alias{num2cut}
\title{Computing breaks for make a histogram of a continuous variable}
\usage{
num2cut(x)
}
\arguments{
\item{x}{A continuous variables}
}
\value{
A list contains a factor and a numeric vector
}
\description{
Computing breaks for make a histogram of a continuous variable
}

