% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPoints.R
\name{num2factorDf}
\alias{num2factorDf}
\title{Make numeric column of a data.frame to factor}
\usage{
num2factorDf(data, colnames, maxfactorno = 6)
}
\arguments{
\item{data}{a data.frame}

\item{colnames}{Column names to be converted}

\item{maxfactorno}{maximun unique value of column}
}
\description{
Make numeric column of a data.frame to factor
}

