% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggChoropleth.R
\name{palette2colors}
\alias{palette2colors}
\title{Extract colors from a palette}
\usage{
palette2colors(name, reverse = FALSE)
}
\arguments{
\item{name}{A palette name from the RColorBrewer package}

\item{reverse}{if true, reverse colors}
}
\description{
Extract colors from a palette
}
