% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggCatepillar.R
\name{pastecolon}
\alias{pastecolon}
\title{Paste character vectors separated by colon}
\usage{
pastecolon(...)
}
\arguments{
\item{...}{Arguments passed on to paste()}
}
\description{
Paste character vectors separated by colon
}
