% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggRadar.R
\name{rescale_df}
\alias{rescale_df}
\title{Rescale all numeric variables of a data.frame except grouping variable}
\usage{
rescale_df(data, groupvar = NULL)
}
\arguments{
\item{data}{A data.frame}

\item{groupvar}{A column name used as a grouping variable}
}
\value{
A rescaled data.frame
}
\description{
Rescale all numeric variables of a data.frame except grouping variable
}
