% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rose}
\alias{rose}
\title{Rose sales among 7 groups in a year}
\format{
An object of class \code{data.frame} with 84 rows and 3 columns.
}
\usage{
rose
}
\description{
A phony dataset representing rose sales
}
\details{
@format A data.frame with 84 rows and 3 columns
 \describe{
    \item{group}{group A to G}
    \item{Month}{Month 1 to 12}
    \item{value}{Rose sales amount}
 }
}
\keyword{datasets}
