% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPieDonut.R
\name{subcolors}
\alias{subcolors}
\title{Make a subcolors according to the mainCol}
\usage{
subcolors(.dta, main, mainCol)
}
\arguments{
\item{.dta}{A data.frame}

\item{main}{A character string of column name used as a main variable}

\item{mainCol}{A main color}
}
\description{
Make a subcolors according to the mainCol
}

