% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Summarize a continuous variable by groups with mean, sd and SE}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, conf.interval = 0.95,
  na.rm = TRUE, .drop = TRUE)
}
\arguments{
\item{data}{A data.frame}

\item{measurevar}{A name of variable to measure a mean and sd}

\item{groupvars}{Name(s) of variable used as a grouping variables}

\item{conf.interval}{confidence interval}

\item{na.rm}{A logical value indicating whether or not remove NA values}

\item{.drop}{should combinations of variables that do not appear in the input data be preserved (FALSE) or dropped (TRUE, default)}
}
\value{
A data.frame summarized a continuous variable by groups with mean, sd and SE
}
\description{
Summarize a continuous variable by groups with mean, sd and SE
}

