% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{taco}
\alias{taco}
\title{Taco ratings by age group}
\format{An object of class \code{data.frame} with 136 rows and 4 columns.}
\usage{
taco
}
\description{
Taco ratings by ShellType, AgeGroup and Filling
source: \href{http://rikturr.com/blog/communicating-experimental-results-with-r/}{Communicating experiment results with R}
}
\details{
@format A data.frame with 136 rows and 4 columns
 \describe{
    \item{ShellType}{Hard or Soft}
    \item{Fillings}{Fillings of taco}
    \item{AgeGroup}{AgeGroup One of the c("<13","13-20","21-39","40+",)}
    \item{Rating}{A numeric. Rating of taco}
 }
}
\keyword{datasets}
