% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_clean.R
\name{theme_clean}
\alias{theme_clean}
\title{Clean theme for PieDonut plot}
\usage{
theme_clean(base_size = 12)
}
\arguments{
\item{base_size}{An integer, default 12.}
}
\description{
Clean theme for PieDonut plot
}
