% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPoints.R
\name{unselectNumeric}
\alias{unselectNumeric}
\title{Unselect numeric column of a data.frame}
\usage{
unselectNumeric(data, colnames, maxfactorno = 6)
}
\arguments{
\item{data}{a data.frame}

\item{colnames}{Column names to be converted}

\item{maxfactorno}{maximun unique value of column}
}
\description{
Unselect numeric column of a data.frame
}
