% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggChoropleth.R
\name{ggChoropleth}
\alias{ggChoropleth}
\title{Draw an interactive choropleth map}
\usage{
ggChoropleth(
  data,
  mapping,
  map,
  palette = "OrRd",
  reverse = FALSE,
  color = "grey50",
  title = "",
  digits = 1,
  interactive = FALSE,
  ...
)
}
\arguments{
\item{data}{a data.frame}

\item{mapping}{Set of aesthetic mappings created by aes or aes_. Passed on geom_map_interactive. Required mappings are map_id and fill. Possible mapping is facet.}

\item{map}{a map maybe a result of map_data()}

\item{palette}{A palette name used for discrete fill var, Default value is "OrRd"}

\item{reverse}{If true, reverse palette colors}

\item{color}{A name of color of polygon, Default value is "grey50"}

\item{title}{A title}

\item{digits}{An integer indicating the number of decimal places}

\item{interactive}{Logical. If positive an interactive map will be made}

\item{...}{other arguments passed on to geom_map_interactive}
}
\description{
Draw an interactive choropleth map
}
\examples{
#crimes <- data.frame(state = tolower(rownames(USArrests)), USArrests)
#require(ggplot2)
#require(ggiraph)
#require(maps)
#require(mapproj)
#require(reshape2)
#require(RColorBrewer)
#states_map <- map_data("state")
#ggChoropleth(crimes,aes(fill=Murder,map_id=state),map=states_map,interactive=TRUE)
#ggChoropleth(crimes,aes(fill=c(Murder,Rape),map_id=state),map=states_map,interactive=TRUE)
#ggChoropleth(crimes,aes(map_id=state),map=states_map,palette="OrRd",interactive=TRUE)
}
