% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggPredict.R
\name{ggPredict}
\alias{ggPredict}
\title{Visualize the multiple linear/logistic regression model with or without interaction}
\usage{
ggPredict(x, ...)
}
\arguments{
\item{x}{Object to ggPredict}

\item{...}{additional arguments passed to the generic function}
}
\description{
Visualize the multiple linear/logistic regression model with or without interaction
}
\examples{
require(ggplot2)
require(ggiraph)
require(moonBook)
require(TH.data)
ggPredict(radial,aes(y=NTAV,x=age,color=sex),method="lm",se=TRUE)
fit=lm(NTAV~age*DM,data=radial)
ggPredict(fit,point=TRUE,interactive=TRUE)
fit1=lm(NTAV~age*weight,data=radial)
ggPredict(fit1,interactive=TRUE,point=TRUE)
fit2=glm(cens~pnodes*horTh,data=GBSG2)
ggPredict(fit2,se=TRUE,interactive=TRUE)
fit3=glm(cens~pnodes*age,data=GBSG2)
ggPredict(fit3,colorn=100,interactive=TRUE)
fit4=glm(cens~pnodes*age*horTh,data=GBSG2)
ggPredict(fit4,interactive=TRUE)
}

